Implement CRUD Routes
Index, Show, New and Create routes implemented for Task model
Create two transformation CLASS methods to convert enum arrays
into hashes for html form select/options.
Have realised one enum may actually be better as it's own model; namely,
routines because it is not enough to simply no it is a recuring routine.
We need to know which days of the week it recurs on. Is it everyday,
every other day or simply every Monday etc.? But the enum for task_type
is valid!