Fix filter by routine
Fixed the issue with the filter by routine not working and when it did
work it wasn't working as expected. Fixed the reject_if proc to be
more specific a subtask is not a subtask if it doesn't have a title.
I've also fixed the set_initial_order method to be more specific to
the task it belongs to. It will now set the order by user_id and max
order. This seems to have fixed the issue of random order on
task creation when view in the index page.