Rework Stimulus controllers - Add inheritance
Created a new branch 'tsj-simulus-division-of-labour' and started to
commit. Created an AbstractTask class, then subclassed both Task and
Subtask classes to simpolify the code. Also added a new controller,
TaskManager, to handle the backend updates and will handle the
triggering of the reordering of the ordinal fields of the tasks and
subtasks and will possibly encapsulate the drag-n-drop functionality,
be that dragular or sortable or some other library, if they facilitate
nested drag-n-drop functionality.