Allow Content Editable & Fix Update Blocking Issue
Both Title and Description are now editable in situ rather than having
to click the 'Edit' button. This is a big improvement for UX.The issue
with the update blocking is now fixed. Seems to have been caused by the
ordinal type; casting it to integer before validation seems to have
fixed it.