Implement the 'recurs on' days functionality
Added store_accessor to :recurs_on so I can access the json object via
the day names. Then added to the form and removed the 'recurs on' input.
Then hid the checkboxes and styled the labels and added a bit of js to
toggle the labels depending on wether the user chooses a routine or not.
Also gave the stimulus controllers a clean up.