Complete routine tasks cloning for next occurence
Also, added a new task types for Stoic Exercises, Stoic Discipline
and Spiritual Development. Only task_type stoic_exercise will be
anylised by AI! Will create a read only journal view for this type.

Changed the default scope to include only active tasks and to order by
start date. Will probably remove the orderinals field for tasks/cards as
it makes more sense to order them by time but will keep it for subtasks.

Fixed a bug I didn't know I had with set_default_start_date method. It
was setting the start date to the current time regardless of whether it
was nil or not. Now it will only set it to the current time if it's nil.

Changed the schema for the tasks table json (recurs_on) field to use 1
and 0 instead of true and false. As the frontend uses them as booleans.