Implement reorder on task date
You can now drag 'n' drop the task and the times (dates) will reorder
automatically and update. I have also added a flatpickr date picker to
the start time field. There was a bug when you dragged the task everything would update etc. But if you opened the datepicker it would be still set at the old time. So if you clicked on it then of it it would resset the time to it's original value. I've fixed that. I turned out to be an issue related to the sortable dragables disconnect/connect callback methods. The fix was essentially that I had to disconnect the flatpickr instances before the reorder and then reconnect them afterwards. The only thing is this is done in the sortable controller. So much for Loose-coupling! I've also removed the ordinals from task cards as they're only needed on subtask cards. I need to trigger an automatic sort, now, when the user changes the time on a task.