Implement task list sort by date after datepicker
I think task lists are now fully functional - famous last words! If you
drag a task to the top of the list, or wherever, the times and under the
hood date/time will reorder themselves. If you click on the time of a
task and choose another time the list will reorder itself to compensate
for the change - NB. You CANNOT Sort by Date-as-String. The comparison
works, which only confused matters but the sort function will not sort
(convert to Date Object!). Or maybe this is just a chrome thing. I
have introduced a new bug. When I clicked on the submit button to create
a new task the form was not removed. I must have pressed it about 5
times as when I gave up and refreshed the page. 5 new idential tasks
appeared. All with the same title - Futue te Ipsum - ironically! I'll
have a look at this tommorrow.