Fix Task form/turbo-frame bug/error
When is :ok not :ok!? It was an easy fix once I worked out what it was.
When fixing the error message I was getting when I moved the ajax/fetch
request to the task controller I had to change the response status in
the backend from :see_other to :ok. And, so, I thought for consistency
I should do the for all responses in the task controller. No, no, no! I
would apear to return zilch when you add :ok to an html response!!! I
found this out after I'd undone a lot of work on the task and index
templates. Because it was a hotwire/turbo-frame isssue I thought I had
caused it with all the stimulus controller data- attributes. But, no,
it wasn't. I wasn't :ok at all Yes it was!!! :-D