Fix profile page sortable
Fixed the profile page sortables. The code was out of date compared to
the task index sortable.

Wrapped the task/subtask controls with a
dropdown menu. Much better with only one button on view now. Added the
confirm date plugin to flatpickr which I thought was a bit of a misnomer
as all it does is add a button to close the datepicker. I had to
customise it to only accept the new date/time when a user clicks the
button. Also had to cutomise the button with markup as the default had
the icon and text butting up against each other.

And installed the material_green theme from flatpickr. Hopefully the
Task/Subtasks are now complete, for now. As, I might totally remove all
of the form/turbo-frame functionality from the UI. Just use TaskCards
and update the backend via ajax/fetch requests as I'm already doing that
for edits why not creation aswell? Probably still use turbo-streams for
deletions, though!