Change Task and Subtask models
Removed Due Date field from both Task and Subtask
as it didn't make sense to have it. Instead I've
added an End Date field to both Task and Subtask
and a duration field in order to beable to monitor
the progress of the task. And the user will be
able to hone their guestimation skills!

I've also added a field for AI generated feedback
to both Task and Subtask. That is if it's possible
to get free and sensible feedback from AI.

Have also started to add a custom domReady
function to my stimulus connect() functions and in
any [identifier]ValueChanged() callbacks to ensure
that there is no annoying cascade of events at
startup.